JobLogAction.java
